test: fix test scripts test: updated AMM test test: https://linear.app/liquality/issue/LIQ-902/update-tests-mocha-settings
