+++ tests/e2e/rsk_tests/10_fastbtc_swap.spec.js
@@ -37,7 +37,6 @@ if (process.env.NODE_ENV === 'mainnet') {
       await overviewPage.HasOverviewPageLoaded(page)
     })
     after(async () => {
-      await page.close()
       await browser.close()
     })
